requirements.txt
streamlit
networkx
pyvis
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
pypdf
numpy
groq
python-dotenv
PyYAML
duckdb
